feat: DB에 데이터 타입 변경 변경

+++ src/main/java/com/chan/stock_portfolio_backtest_api/domain/CalcStockPrice.java
@@ -19,8 +19,8 @@ public class CalcStockPrice {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    private Integer price;
-    private float monthlyRor;
+    private Float price;
+    private Float monthlyRor;
     private LocalDate baseDate;
 
     @JsonIgnore
